CMakeLists.txt executables are made more-or-less properly
 - run_tests
 - run_benchmarks
- removes macros gating entire source files
- renamed main.cpp run_tests.cpp
