+++ src/run_tests.cpp
@@ -1,0 +1,16 @@
+#include <cstdio>
+
+void fixed_point_test();
+void fixed_point_utils_test();
+void proposal_test();
+
+int main(int, char *[])
+{
+	fixed_point_test();
+	fixed_point_utils_test();
+	proposal_test();
+
+	std::puts("Tests complete.");
+
+	return 0;
+}
